Fix issue when response contains existing class
If we receive a response like 

<Response><timestamp>1394617345</timestamp><num_bytes>6625</num_bytes><Trip>....

and we have an existing class called Trip, the process could crash while trying to create the Trip object. The code in TravelObj#new tries to get and crate constants that belongs to a module, but it does it inheriting. The change just avoids inheriting.